refactor: extract context menu creation to a factory class
