+++ src/main/java/core/packetproxy/gui/GUIHistory.java
@@ -24,8 +24,6 @@ import java.awt.Component;
 import java.awt.Dimension;
 import java.awt.Rectangle;
 import java.awt.Toolkit;
-import java.awt.datatransfer.Clipboard;
-import java.awt.datatransfer.StringSelection;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.ComponentAdapter;
@@ -70,14 +68,10 @@ import javax.swing.event.TableModelListener;
 import javax.swing.table.DefaultTableModel;
 import javax.swing.table.TableCellRenderer;
 import javax.swing.table.TableRowSorter;
-import org.apache.commons.io.FileUtils;
 import packetproxy.common.FilterTextParser;
 import packetproxy.common.FontManager;
 import packetproxy.common.I18nString;
 import packetproxy.common.Utils;
-import packetproxy.controller.ResendController;
-import packetproxy.http.HeaderField;
-import packetproxy.http.Http;
 import packetproxy.model.Database;
 import packetproxy.model.Database.DatabaseMessage;
 import packetproxy.model.Filters;
@@ -85,7 +79,6 @@ import packetproxy.model.OptionTableModel;
 import packetproxy.model.Packet;
 import packetproxy.model.Packets;
 import packetproxy.model.ResenderPackets;
-import packetproxy.util.CharSetUtility;
 
 public class GUIHistory implements PropertyChangeListener {
 
@@ -438,381 +431,13 @@ public class GUIHistory implements PropertyChangeListener {
 		sorter.toggleSortOrder(14); /* 14 is 'group' column */
 		table.setRowSorter(sorter);
 
-		int mask_key = Toolkit.getDefaultToolkit().getMenuShortcutKeyMaskEx();
-		menu = new JPopupMenu();
-
-		JMenuItem send = createMenuItem("send", KeyEvent.VK_S, KeyStroke.getKeyStroke(KeyEvent.VK_S, mask_key),
-				new ActionListener() {
-
-					public void actionPerformed(ActionEvent actionEvent) {
-						try {
-
-							int id = GUIHistory.getInstance().getSelectedPacketId();
-							Packet packet = Packets.getInstance().query(id);
-							byte[] data = GUIPacket.getInstance().getData();
-							if (packet == null) {
-
-								return;
-							}
-							ResendController.getInstance().resend(packet.getOneShotPacket(data));
-							packet.setResend();
-							Packets.getInstance().update(packet);
-							GUIHistory.getInstance().updateRequestOne(GUIHistory.getInstance().getSelectedPacketId());
-						} catch (Exception e1) {
-
-							errWithStackTrace(e1);
-						}
-					}
-				});
-
-		JMenuItem sendToResender = createMenuItem("send to Resender", KeyEvent.VK_R,
-				KeyStroke.getKeyStroke(KeyEvent.VK_R, mask_key), new ActionListener() {
-
-					public void actionPerformed(ActionEvent actionEvent) {
-						try {
-
-							Packet packet = gui_packet.getPacket();
-							packet.setResend();
-							Packets.getInstance().update(packet);
-							if (packet.getModifiedData().length == 0) { // dropしたパケットの場合
-
-								GUIResender.getInstance().addResends(packet.getOneShotFromDecodedData());
-							} else {
-
-								GUIResender.getInstance().addResends(packet.getOneShotFromModifiedData());
-							}
-							GUIHistory.getInstance().updateRequestOne(GUIHistory.getInstance().getSelectedPacketId());
-						} catch (Exception e1) {
-
-							errWithStackTrace(e1);
-						}
-					}
-				});
-
-		JMenuItem copyAll = createMenuItem("copy Method + URL + Body", KeyEvent.VK_M,
-				KeyStroke.getKeyStroke(KeyEvent.VK_M, mask_key), new ActionListener() {
-
-					public void actionPerformed(ActionEvent actionEvent) {
-						try {
-
-							Packet packet = gui_packet.getPacket();
-							Http http = Http.create(packet.getDecodedData());
-							CharSetUtility charsetutil = CharSetUtility.getInstance();
-							if (charsetutil.isAuto()) {
-
-								charsetutil.setGuessedCharSet(http.getBody());
-							}
-							String copyData = http.getMethod() + "\t"
-									+ http.getURL(packet.getServerPort(), packet.getUseSSL()) + "\t"
-									+ new String(http.getBody(), charsetutil.getCharSet());
-							Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
-							StringSelection selection = new StringSelection(copyData);
-							clipboard.setContents(selection, selection);
-						} catch (Exception e1) {
-
-							errWithStackTrace(e1);
-						}
-					}
-				});
-
-		JMenuItem copy = createMenuItem("copy URL", KeyEvent.VK_Y, KeyStroke.getKeyStroke(KeyEvent.VK_Y, mask_key),
-				new ActionListener() {
-
-					public void actionPerformed(ActionEvent actionEvent) {
-						try {
-
-							int id = GUIHistory.getInstance().getSelectedPacketId();
-							Packet packet = Packets.getInstance().query(id);
-							Http http = Http.create(packet.getDecodedData());
-							String url = http.getURL(packet.getServerPort(), packet.getUseSSL());
-							Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
-							StringSelection selection = new StringSelection(url);
-							clipboard.setContents(selection, selection);
-
-						} catch (Exception e1) {
-
-							errWithStackTrace(e1);
-						}
-					}
-				});
-
-		JMenuItem bulkSender = createMenuItem("send to Bulk Sender", -1, null, new ActionListener() {
-
-			public void actionPerformed(ActionEvent actionEvent) {
-				try {
-
-					Packet packet = gui_packet.getPacket();
-					if (packet.getModifiedData().length == 0) { // dropしたパケットの場合
-
-						GUIBulkSender.getInstance().add(packet.getOneShotFromDecodedData(), packet.getId());
-					} else {
-
-						GUIBulkSender.getInstance().add(packet.getOneShotFromModifiedData(), packet.getId());
-					}
-				} catch (Exception e) {
-
-					errWithStackTrace(e);
-				}
-			}
-		});
-
-		JMenuItem saveAll = createMenuItem("save all data to file", -1, null, new ActionListener() {
-
-			public void actionPerformed(ActionEvent actionEvent) {
-				WriteFileChooserWrapper filechooser = new WriteFileChooserWrapper(owner, "dat", "packet.dat");
-				filechooser.addFileChooserListener(new WriteFileChooserWrapper.FileChooserListener() {
-
-					@Override
-					public void onApproved(File file, String extension) {
-						try {
-
-							byte[] data = gui_packet.getPacket().getReceivedData();
-							FileUtils.writeByteArrayToFile(file, data);
-							JOptionPane.showMessageDialog(owner, String.format("%sに保存しました！", file.getPath()));
-						} catch (Exception e1) {
-
-							errWithStackTrace(e1);
-							JOptionPane.showMessageDialog(null, "データの保存に失敗しました。");
-						}
-						dialogOnce = false;
-					}
-
-					@Override
-					public void onCanceled() {
-					}
-
-					@Override
-					public void onError() {
-						JOptionPane.showMessageDialog(null, "データの保存に失敗しました。");
-					}
-				});
-				filechooser.showSaveDialog();
-			}
-		});
-
-		JMenuItem saveHttpBody = createMenuItem("save HTTP body to file", -1, null, new ActionListener() {
-
-			public void actionPerformed(ActionEvent actionEvent) {
-				WriteFileChooserWrapper filechooser = new WriteFileChooserWrapper(owner, "dat", "body.dat");
-				filechooser.addFileChooserListener(new WriteFileChooserWrapper.FileChooserListener() {
-
-					@Override
-					public void onApproved(File file, String extension) {
-						try {
-
-							Http http = Http.create(gui_packet.getPacket().getDecodedData());
-							byte[] data = http.getBody();
-							FileUtils.writeByteArrayToFile(file, data);
-							JOptionPane.showMessageDialog(owner, String.format("%sに保存しました！", file.getPath()));
-						} catch (Exception e1) {
-
-							errWithStackTrace(e1);
-							JOptionPane.showMessageDialog(null, "データの保存に失敗しました。");
-						}
-						dialogOnce = false;
-					}
-
-					@Override
-					public void onCanceled() {
-					}
-
-					@Override
-					public void onError() {
-						JOptionPane.showMessageDialog(null, "データの保存に失敗しました。");
-					}
-				});
-				filechooser.showSaveDialog();
-			}
-		});
-
-		JMenuItem addColorG = createMenuItem("add color (green)", -1, null, new ActionListener() {
-
-			public void actionPerformed(ActionEvent actionEvent) {
-				try {
-
-					int[] selected_rows = table.getSelectedRows();
-					for (int i = 0; i < selected_rows.length; i++) {
-
-						Integer id = (Integer) table.getValueAt(selected_rows[i], 0);
-						colorManager.add(id, packetColorGreen);
-						Packet packet = Packets.getInstance().query(id);
-						packet.setColor("green");
-						Packets.getInstance().update(packet);
-					}
-				} catch (Exception e) {
-
-					errWithStackTrace(e);
-				}
-			}
-		});
-
-		JMenuItem addColorB = createMenuItem("add color (brown)", -1, null, new ActionListener() {
-
-			public void actionPerformed(ActionEvent actionEvent) {
-				try {
-
-					int[] selected_rows = table.getSelectedRows();
-					for (int i = 0; i < selected_rows.length; i++) {
-
-						Integer id = (Integer) table.getValueAt(selected_rows[i], 0);
-						colorManager.add(id, packetColorBrown);
-						Packet packet = Packets.getInstance().query(id);
-						packet.setColor("brown");
-						Packets.getInstance().update(packet);
-					}
-				} catch (Exception e) {
-
-					errWithStackTrace(e);
-				}
-			}
-		});
-
-		JMenuItem addColorY = createMenuItem("add color (yellow)", -1, null, new ActionListener() {
-
-			public void actionPerformed(ActionEvent actionEvent) {
-				try {
-
-					int[] selected_rows = table.getSelectedRows();
-					for (int i = 0; i < selected_rows.length; i++) {
-
-						Integer id = (Integer) table.getValueAt(selected_rows[i], 0);
-						colorManager.add(id, packetColorYellow);
-						Packet packet = Packets.getInstance().query(id);
-						packet.setColor("yellow");
-						Packets.getInstance().update(packet);
-					}
-				} catch (Exception e) {
-
-					errWithStackTrace(e);
-				}
-			}
-		});
-
-		JMenuItem clearColor = createMenuItem("clear color", -1, null, new ActionListener() {
-
-			public void actionPerformed(ActionEvent actionEvent) {
-				try {
-
-					int[] selected_rows = table.getSelectedRows();
-					for (int i = 0; i < selected_rows.length; i++) {
-
-						Integer id = (Integer) table.getValueAt(selected_rows[i], 0);
-						colorManager.clear(id);
-					}
-				} catch (Exception e) {
-
-					errWithStackTrace(e);
-				}
-			}
-		});
-
-		JMenuItem delete_selected_items = createMenuItem("delete selected items", -1, null, new ActionListener() {
-
-			public void actionPerformed(ActionEvent actionEvent) {
-				try {
-
-					int[] selected_rows = table.getSelectedRows();
-					for (int i = 0; i < selected_rows.length; i++) {
-
-						Integer id = (Integer) table.getValueAt(selected_rows[i], 0);
-						colorManager.clear(id);
-						packets.delete(packets.query(id));
-					}
-					updateAll();
-				} catch (Exception e2) {
-
-					errWithStackTrace(e2);
-				}
-			}
-		});
-
-		JMenuItem delete_all = createMenuItem("delete all items", -1, null, new ActionListener() {
-
-			public void actionPerformed(ActionEvent actionEvent) {
-				try {
-
-					for (int i = 0; i < table.getRowCount(); ++i) {
-
-						Integer id = (Integer) table.getValueAt(i, 0);
-						colorManager.clear(id);
-					}
-					packets.deleteAll();
-					updateAll();
-				} catch (Exception e2) {
-
-					errWithStackTrace(e2);
-				}
-			}
-		});
-
-		/*
-		 * Copy HTTP request as curl command:
-		 * $ curl 'http://example.com' -X POST -H 'Cookie: hoge=huga; foo=bar' -H ...
-		 *
-		 * TODO: support --data-binary
-		 */
-		JMenuItem copyAsCurl = createMenuItem("copy as curl", -1, null, new ActionListener() {
-
-			@Override
-			public void actionPerformed(ActionEvent actionEvent) {
-				try {
-
-					Http http = Http.create(gui_packet.getPacket().getDecodedData());
-					List<HeaderField> headerFields = http.getHeader().getFields();
-					ArrayList<String> commandList = new ArrayList<>();
-					commandList.add("curl");
-
-					// 'http://example.com'
-					String url = http.getURL(gui_packet.getPacket().getServerPort(),
-							gui_packet.getPacket().getUseSSL());
-					commandList.add(String.format("'%s'", url));
-
-					// -X POST
-					commandList.add("-X");
-					commandList.add(http.getMethod());
-
-					// -H 'Cookie: hoge=huga; foo=bar' -H ...
-					for (HeaderField hf : headerFields) {
-
-						commandList.add("-H");
-						commandList.add(String.format("'%s: %s'", hf.getName(), hf.getValue()));
-					}
-
-					// --data 'id=hoge&password=huga'
-					String body = new String(http.getBody());
-					// if body is not empty
-					if (body.trim().length() > 0) {
-
-						commandList.add("--data");
-						commandList.add(String.format("'%s'", body));
-					}
-
-					commandList.add("--compressed");
-					StringSelection command = new StringSelection(String.join(" ", commandList));
-
-					Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
-					clipboard.setContents(command, command);
-				} catch (Exception e) {
-
-					errWithStackTrace(e);
-				}
-			}
-		});
-
-		menu.add(send);
-		menu.add(sendToResender);
-		menu.add(copyAll);
-		menu.add(copy);
-		menu.add(bulkSender);
-		menu.add(saveAll);
-		menu.add(saveHttpBody);
-		menu.add(addColorG);
-		menu.add(addColorB);
-		menu.add(addColorY);
-		menu.add(clearColor);
-		menu.add(copyAsCurl);
-		menu.add(delete_selected_items);
-		menu.add(delete_all);
+		GUIHistoryContextMenuFactory.Handles handles = GUIHistoryContextMenuFactory.build(this, owner, table,
+				gui_packet, packets, colorManager, packetColorGreen, packetColorBrown, packetColorYellow);
+		menu = handles.getMenu();
+		JMenuItem send = handles.getSend();
+		JMenuItem sendToResender = handles.getSendToResender();
+		JMenuItem copy = handles.getCopy();
+		JMenuItem copyAll = handles.getCopyAll();
 
 		table.addKeyListener(new KeyAdapter() {
